Congratulations on making it this far.
You are seeking answers.  You want to know how this all began...

July 6, 2021: It started on a day like any other.  I was sitting in a
vegan cafe in Los Angeles that serves breakfast, burgers and Mexican
food.  As we sat there sipping matcha bubble tea oat "milk" lattes and
eating buffalo cauliflower "wings", we conceived the idea for Stargate
and wrote all of the code for this first commit on the back of a few
thousand napkins.  Special thanks to the ghost of Kurt Cobain, and the
race of lizard people living under the city.

Realzing that we were on to something big, I quickly drove my Model 3
to Joshua Tree, climbed to the top of Giant Rock.  For 9 days and
I sat on top of Giant Rock with my laptop, carefully transcribing the
contents of those napkins into this Git repo.

The rest, as they say, is history.