Fix a crash
When switching from a track with an SG EQ, to a track with a Sampler1,
the long load time of Sampler1 causes events to pile up, overflow and
SEGFAULT in the SGEQ spectrum analyzer.  I need to go back and come up
with a better fix, but for now disabling the old rack UI messages before
loading the new rack is an adequate solution.