Normalize SG Channel volume against the pan law setting
This will affect existing projects.  Prior to this commit, the volume of
the channel would be affected by inserting an SG Channel by -3.0dB, or
whatever you changed the pan law setting to.  This will no longer
happen, but you may need to go back and add +3dB of gain to any mixer
channel with an SG Channel plugin.