Fix crashes on Linux.
Because when you are a security engineer, crashing somebody else's app
or adversely affecting the performance of it is a small price to pay to
mitigate what a hacker can do to a system they already have control of.
Don't even give users the option of being "insecure", it's the
security engineer's way.  So disable any attempts to optimize the
scheduler for latency, and just take what the distros give us.