Update licensing information for compliance
The GNU General Public License requires that a copy of the license
is distributed with the sources. This also helps code license detection
systems determine that the software is actually licensed under the GPLv3.

The "license statement" is now in the README instead of the LICENSE
file, and the LICENSE file now contains the actual license.