Wave Editor: Fix potential engine crash when changing files
I had been trying to hunt down this bug for months.  It was what I
thought it was:  A compiler bug.  I decided to harden the engine for
stack smashing, assuming that could be it and explain why this is so
hard to debug.  Swapping a single instance of sprintf for my wrapped
version of snprintf fixed the issue, but there was no flaw in the code
revealed by this, no buffer overflow to fix, just switching one
particular invocation to a different *printf function fixed it.