Work around stupid MacOS limitations
UDP is limited to 64kb of data sent in one go.  Which is already
very limiting.  But Apple arbitrarily decided that is should really
be limited to 9216 bytes, because 64kb is obviously too much.  Which
broke the spectrum analyzer, and blocked any future commmunication with
the UI once the spectrum analyzer is activated.  We probably should
come up with a special IPC mechanism just for MacOS to get around this.
As a stop-gap measure:

1. Do a preflight check for the UDP max size, warn the user to change it
2. Timeouts on the UI response so that the socket thread does not hang
3. Documentation updates